Remove the old implementation of the decode / encode payloads.
The new versions are based on the same semantics as Node.js. They are
actually tied to the transports and abstract that wire protocol from the
message queues.

Also, fix the JSONP protocol encoding details.

Most importantly, implemented the optimization with the multi packets
reply for long-polling protocols, which will make the replies way more
efficient.