Added reset_acl()
Renamed call_method to call_method_with_acl

call_method is now a function that checks the argspecs for a user-defined
method, and if the first parameter is named 'packet' it sends in the full
packet dict, for introspection by the developer.

Renamed recv_initialize() to initialize() and documented it, since we do not
receive an initialize per se, so it is a high level function, that gets
passed the packet that triggers that, but is instructed not to act upon it,
since it is an unsecured method.