Fix nickname logic, add disconnect logic.
The nickname stuff wasn't working correctly before because it was setting self.environ['nicknames'], but self.environ is new for each new socket connection.  I'm unsure of the right way to do this, so I'd like some feedback here.  I maintain state by using a request object which I pass into the manage function.

What I'm assuming, based on reading the docs, is that a new Namespace instance is created /per/ socket connection, but a new instance *isn't* created between multiple requests on the *same* socket in the case of a long poll.  In this case, it makes sense to track the nicknames in the request object, right?